requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dateparser==1.0.0
Django==3.1.3
django-autoslug==1.9.8
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
idna==2.10
install==1.3.4
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==2.0.9
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
r-django-jet==2.0.9
redis==3.5.3
regex==2020.11.13
requests==2.25.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
stories==4.0.0
tzlocal==2.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
